Rename dakota_utils module utils
